+++ js/data/constants.js
@@ -1,0 +1,33 @@
+const PHOTO_COUNT = 25;
+const AVATAR_COUNT = 6;
+const MAX_COMMENT_COUNT = 5;
+const MIN_LIKE_COUNT = 15;
+const MAX_LIKE_COUNT = 200;
+
+const DESCRIPTIONS = [
+  'Закат на пляже',
+  'Горы зимой',
+  'Городская улица ночь',
+  'Лесная тропинка',
+  'Весенний цветок',
+  'Осенний парк',
+  'Морской пейзаж',
+  'Заброшенный дом',
+  'Старый мост',
+  'Туманный лес',
+  'Рассвет в горах'
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const NAMES = ['Артём', 'Иван', 'Ольга', 'Никита', 'Мария', 'Алексей', 'Анна', 'Дмитрий', 'Екатерина', 'Сергей'];
+
+
+export { PHOTO_COUNT, AVATAR_COUNT, MAX_COMMENT_COUNT, MIN_LIKE_COUNT, MAX_LIKE_COUNT, DESCRIPTIONS, MESSAGES, NAMES };
